Divided Wishlist UI to components
